fix: Remove the possibility to circumvent event delegation
Commit 0c95dd2 fixed a bug where it was possible to click
the deck (space between cards, or the edge) and it would
add the 'open' class to the deck element itself, thus revealing
all cards. However, that fix introduced another bug which allowed
the user to click one card, then keep the mouse pressed and move it
to another card, then release. This also caused all symbols to show up.
This is now fixed. 'pointer-events' is not a good way to deal with
event delegation in this particular case.